update vue3-vite-ts project
